Plugin - Output instance and listening destination input changes
We now create an Output instance in the run method so it can be modified while using the plugin. This allows to edit the destination and to add or remove fields to the disabled and hidden lists.
Adding and event listener on the destination input so we can modify the directory attribute of the Output instance.
Note that we now use lambdas within connect methods so we can pass args